fix: fix a bug related to delete account api that used to lead to an error
